refact : Lines에 있는 기능(findExpensiveFare)을 EffecinetLiens로 이동
