+++ src/main/java/nextstep/subway/path/domain/FareCalculator.java
@@ -11,7 +11,7 @@ public class FareCalculator {
     public static Money calcFare(LoginMember loginMember, ShortestDistance shortestDistance) {
         Money money = calcDistance(shortestDistance.shortestDistance());
         money = calcAge(loginMember, money);
-        money = calcLines(shortestDistance.usedLines(), money);
+        money = calcLines(new EfficientLines(shortestDistance.usedLines()), money);
 
         return money;
     }
