+++ src/main/java/nextstep/subway/path/domain/LinePremiumPolicy.java
@@ -4,7 +4,7 @@ import nextstep.subway.line.domain.Lines;
 import nextstep.subway.wrapped.Money;
 
 public interface LinePremiumPolicy {
-    Money calcFare(Lines lines, Money money);
+    Money calcFare(EfficientLines lines, Money money);
 
-    boolean isSupport(Lines lines);
+    boolean isSupport(EfficientLines lines);
 }
