+++ src/main/java/nextstep/subway/path/domain/line/DefaultLinePremiumPolicy.java
@@ -1,17 +1,17 @@
 package nextstep.subway.path.domain.line;
 
-import nextstep.subway.line.domain.Lines;
+import nextstep.subway.path.domain.EfficientLines;
 import nextstep.subway.path.domain.LinePremiumPolicy;
 import nextstep.subway.wrapped.Money;
 
 public class DefaultLinePremiumPolicy implements LinePremiumPolicy {
     @Override
-    public Money calcFare(Lines lines, Money money) {
+    public Money calcFare(EfficientLines lines, Money money) {
         return money.plus(lines.findExpensiveFare());
     }
 
     @Override
-    public boolean isSupport(Lines lines) {
+    public boolean isSupport(EfficientLines lines) {
         return true;
     }
 }
